autologistic model in a spatial setting
Data comes from a study in agriculture. The primary aim is the study of the spatial structure of the Carex Arenaria in a field with different species of plants. We also want to assess if the abundance of other plants helps in predicting the presence of Carex Arenaria in a certain location.